initial changes for analysis_utils refactor
